Update user defined class
This change updates the user defined class to write a random string file
out based on a seed of the user's passwords. This is also pseudo-random
based on the fqdn which means it's random but it won't change between
Puppet runs. That allows us to use it deterministically to evaluate if
the user's password needs to be changed so we're not stopping/starting the
service every 30 minutes for every user.

Also fixed some random things here and there and moved the resources that
are static in the user defined type out to the config class. Ordered the config
class before the Snmp::User being evaluated as well.